requirements.txt
asgiref==3.8.1
azure-ai-language-conversations==1.1.0
azure-ai-language-questionanswering==1.1.0
azure-core==1.32.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-environ==0.12.0
idna==3.10
isodate==0.7.2
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
